// damn-master (5)/damn-master/frontend/src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
	ID: { input: string; output: string };
	String: { input: string; output: string };
	Boolean: { input: boolean; output: boolean };
	Int: { input: number; output: number };
	Float: { input: number; output: number };
	/** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
	DateTime: { input: any; output: any };
	/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
	JSON: { input: any; output: any };
	/** The `Money` scalar type represents monetary values and supports signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
	Money: { input: any; output: any };
	/** The `Upload` scalar type represents a file upload. */
	Upload: { input: any; output: any };
};

export type ActiveOrderResult = NoActiveOrderError | Order;

export type AddPaymentToOrderResult =
	| IneligiblePaymentMethodError
	| NoActiveOrderError
	| Order
	| OrderPaymentStateError
	| OrderStateTransitionError
	| PaymentDeclinedError
	| PaymentFailedError;

export type Address = Node & {
	__typename?: 'Address';
	city?: Maybe<Scalars['String']['output']>;
	company?: Maybe<Scalars['String']['output']>;
	country: Country;
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	defaultBillingAddress?: Maybe<Scalars['Boolean']['output']>;
	defaultShippingAddress?: Maybe<Scalars['Boolean']['output']>;
	fullName?: Maybe<Scalars['String']['output']>;
	id: Scalars['ID']['output'];
	phoneNumber?: Maybe<Scalars['String']['output']>;
	postalCode?: Maybe<Scalars['String']['output']>;
	province?: Maybe<Scalars['String']['output']>;
	streetLine1: Scalars['String']['output'];
	streetLine2?: Maybe<Scalars['String']['output']>;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Country = Node &
	Region & {
		__typename?: 'Country';
		code: Scalars['String']['output'];
		createdAt: Scalars['DateTime']['output'];
		customFields?: Maybe<Scalars['JSON']['output']>;
		enabled: Scalars['Boolean']['output'];
		id: Scalars['ID']['output'];
		languageCode: LanguageCode;
		name: Scalars['String']['output'];
		parent?: Maybe<Region>;
		parentId?: Maybe<Scalars['ID']['output']>;
		translations: Array<RegionTranslation>;
		type: Scalars['String']['output'];
		updatedAt: Scalars['DateTime']['output'];
	};
// ...
export type CreateAddressInput = {
	city?: InputMaybe<Scalars['String']['input']>;
	company?: InputMaybe<Scalars['String']['input']>;
	countryCode: Scalars['String']['input'];
	customFields?: InputMaybe<Scalars['JSON']['input']>;
	defaultBillingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	defaultShippingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	fullName?: InputMaybe<Scalars['String']['input']>;
	phoneNumber?: InputMaybe<Scalars['String']['input']>;
	postalCode?: InputMaybe<Scalars['String']['input']>;
	province?: InputMaybe<Scalars['String']['input']>;
	streetLine1: Scalars['String']['input'];
	streetLine2?: InputMaybe<Scalars['String']['input']>;
};
// ...
export const LanguageCode = {
	/** Afrikaans */
	Af: 'af',
	/** Akan */
	Ak: 'ak',
	/** Amharic */
	Am: 'am',
	/** Arabic */
	Ar: 'ar',
	/** Assamese */
	As: 'as',
	/** Azerbaijani */
	Az: 'az',
	/** Belarusian */
	Be: 'be',
	/** Bulgarian */
	Bg: 'bg',
	/** Bambara */
	Bm: 'bm',
	/** Bangla */
	Bn: 'bn',
	/** Tibetan */
	Bo: 'bo',
	/** Breton */
	Br: 'br',
	/** Bosnian */
	Bs: 'bs',
	/** Catalan */
	Ca: 'ca',
	/** Chechen */
	Ce: 'ce',
	/** Corsican */
	Co: 'co',
	/** Czech */
	Cs: 'cs',
	/** Church Slavic */
	Cu: 'cu',
	/** Welsh */
	Cy: 'cy',
	/** Danish */
	Da: 'da',
	/** German */
	De: 'de',
	/** Austrian German */
	DeAt: 'de_AT',
	/** Swiss High German */
	DeCh: 'de_CH',
	/** Dzongkha */
	Dz: 'dz',
	/** Ewe */
	Ee: 'ee',
	/** Greek */
	El: 'el',
	/** English */
	En: 'en',
	/** Australian English */
	EnAu: 'en_AU',
	/** Canadian English */
	EnCa: 'en_CA',
	/** British English */
	EnGb: 'en_GB',
	/** American English */
	EnUs: 'en_US',
	/** Esperanto */
	Eo: 'eo',
	/** Spanish */
	Es: 'es',
	/** European Spanish */
	EsEs: 'es_ES',
	/** Mexican Spanish */
	EsMx: 'es_MX',
	/** Estonian */
	Et: 'et',
	/** Basque */
	Eu: 'eu',
	/** Persian */
	Fa: 'fa',
	/** Dari */
	FaAf: 'fa_AF',
	/** Fulah */
	Ff: 'ff',
	/** Finnish */
	Fi: 'fi',
	/** Faroese */
	Fo: 'fo',
	/** French */
	Fr: 'fr',
	/** Canadian French */
	FrCa: 'fr_CA',
	/** Swiss French */
	FrCh: 'fr_CH',
	/** Western Frisian */
	Fy: 'fy',
	/** Irish */
	Ga: 'ga',
	/** Scottish Gaelic */
	Gd: 'gd',
	/** Galician */
	Gl: 'gl',
	/** Gujarati */
	Gu: 'gu',
	/** Manx */
	Gv: 'gv',
	/** Hausa */
	Ha: 'ha',
	/** Hebrew */
	He: 'he',
	/** Hindi */
	Hi: 'hi',
	/** Croatian */
	Hr: 'hr',
	/** Haitian Creole */
	Ht: 'ht',
	/** Hungarian */
	Hu: 'hu',
	/** Armenian */
	Hy: 'hy',
	/** Interlingua */
	Ia: 'ia',
	/** Indonesian */
	Id: 'id',
	/** Igbo */
	Ig: 'ig',
	/** Sichuan Yi */
	Ii: 'ii',
	/** Icelandic */
	Is: 'is',
	/** Italian */
	It: 'it',
	/** Japanese */
	Ja: 'ja',
	/** Javanese */
	Jv: 'jv',
	/** Georgian */
	Ka: 'ka',
	/** Kikuyu */
	Ki: 'ki',
	/** Kazakh */
	Kk: 'kk',
	/** Kalaallisut */
	Kl: 'kl',
	/** Khmer */
	Km: 'km',
	/** Kannada */
	Kn: 'kn',
	/** Korean */
	Ko: 'ko',
	/** Kashmiri */
	Ks: 'ks',
	/** Kurdish */
	Ku: 'ku',
	/** Cornish */
	Kw: 'kw',
	/** Kyrgyz */
	Ky: 'ky',
	/** Latin */
	La: 'la',
	/** Luxembourgish */
	Lb: 'lb',
	/** Ganda */
	Lg: 'lg',
	/** Lingala */
	Ln: 'ln',
	/** Lao */
	Lo: 'lo',
	/** Lithuanian */
	Lt: 'lt',
	/** Luba-Katanga */
	Lu: 'lu',
	/** Latvian */
	Lv: 'lv',
	/** Malagasy */
	Mg: 'mg',
	/** Maori */
	Mi: 'mi',
	/** Macedonian */
	Mk: 'mk',
	/** Malayalam */
	Ml: 'ml',
	/** Mongolian */
	Mn: 'mn',
	/** Marathi */
	Mr: 'mr',
	/** Malay */
	Ms: 'ms',
	/** Maltese */
	Mt: 'mt',
	/** Burmese */
	My: 'my',
	/** Norwegian Bokmål */
	Nb: 'nb',
	/** North Ndebele */
	Nd: 'nd',
	/** Nepali */
	Ne: 'ne',
	/** Dutch */
	Nl: 'nl',
	/** Flemish */
	NlBe: 'nl_BE',
	/** Norwegian Nynorsk */
	Nn: 'nn',
	/** Nyanja */
	Ny: 'ny',
	/** Oromo */
	Om: 'om',
	/** Odia */
	Or: 'or',
	/** Ossetic */
	Os: 'os',
	/** Punjabi */
	Pa: 'pa',
	/** Polish */
	Pl: 'pl',
	/** Pashto */
	Ps: 'ps',
	/** Portuguese */
	Pt: 'pt',
	/** Brazilian Portuguese */
	PtBr: 'pt_BR',
	/** European Portuguese */
	PtPt: 'pt_PT',
	/** Quechua */
	Qu: 'qu',
	/** Romansh */
	Rm: 'rm',
	/** Rundi */
	Rn: 'rn',
	/** Romanian */
	Ro: 'ro',
	/** Moldavian */
	RoMd: 'ro_MD',
	/** Russian */
	Ru: 'ru',
	/** Kinyarwanda */
	Rw: 'rw',
	/** Sanskrit */
	Sa: 'sa',
	/** Sindhi */
	Sd: 'sd',
	/** Northern Sami */
	Se: 'se',
	/** Sango */
	Sg: 'sg',
	/** Sinhala */
	Si: 'si',
	/** Slovak */
	Sk: 'sk',
	/** Slovenian */
	Sl: 'sl',
	/** Samoan */
	Sm: 'sm',
	/** Shona */
	Sn: 'sn',
	/** Somali */
	So: 'so',
	/** Albanian */
	Sq: 'sq',
	/** Serbian */
	Sr: 'sr',
	/** Southern Sotho */
	St: 'st',
	/** Sundanese */
	Su: 'su',
	/** Swedish */
	Sv: 'sv',
	/** Swahili */
	Sw: 'sw',
	/** Congo Swahili */
	SwCd: 'sw_CD',
	/** Tamil */
	Ta: 'ta',
	/** Telugu */
	Te: 'te',
	/** Tajik */
	Tg: 'tg',
	/** Thai */
	Th: 'th',
	/** Tigrinya */
	Ti: 'ti',
	/** Turkmen */
	Tk: 'tk',
	/** Tongan */
	To: 'to',
	/** Turkish */
	Tr: 'tr',
	/** Tatar */
	Tt: 'tt',
	/** Uyghur */
	Ug: 'ug',
	/** Ukrainian */
	Uk: 'uk',
	/** Urdu */
	Ur: 'ur',
	/** Uzbek */
	Uz: 'uz',
	/** Vietnamese */
	Vi: 'vi',
	/** Volapük */
	Vo: 'vo',
	/** Wolof */
	Wo: 'wo',
	/** Xhosa */
	Xh: 'xh',
	/** Yiddish */
	Yi: 'yi',
	/** Yoruba */
	Yo: 'yo',
	/** Chinese */
	Zh: 'zh',
	/** Simplified Chinese */
	ZhHans: 'zh_Hans',
	/** Traditional Chinese */
	ZhHant: 'zh_Hant',
	/** Zulu */
	Zu: 'zu',
} as const;

export type LanguageCode = (typeof LanguageCode)[keyof typeof LanguageCode];
// ...
export type Node = {
	id: Scalars['ID']['output'];
};
// ...
export type Region = {
	code: Scalars['String']['output'];
	createdAt: Scalars['DateTime']['output'];
	enabled: Scalars['Boolean']['output'];
	id: Scalars['ID']['output'];
	languageCode: LanguageCode;
	name: Scalars['String']['output'];
	parent?: Maybe<Region>;
	parentId?: Maybe<Scalars['ID']['output']>;
	translations: Array<RegionTranslation>;
	type: Scalars['String']['output'];
	updatedAt: Scalars['DateTime']['output'];
};

export type RegionTranslation = {
	__typename?: 'RegionTranslation';
	createdAt: Scalars['DateTime']['output'];
	id: Scalars['ID']['output'];
	languageCode: LanguageCode;
	name: Scalars['String']['output'];
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type UpdateAddressInput = {
	city?: InputMaybe<Scalars['String']['input']>;
	company?: InputMaybe<Scalars['String']['input']>;
	countryCode?: InputMaybe<Scalars['String']['input']>;
	customFields?: InputMaybe<Scalars['JSON']['input']>;
	defaultBillingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	defaultShippingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	fullName?: InputMaybe<Scalars['String']['input']>;
	id: Scalars['ID']['input'];
	phoneNumber?: InputMaybe<Scalars['String']['input']>;
	postalCode?: InputMaybe<Scalars['String']['input']>;
	province?: InputMaybe<Scalars['String']['input']>;
	streetLine1?: InputMaybe<Scalars['String']['input']>;
	streetLine2?: InputMaybe<Scalars['String']['input']>;
};